Separate Game component
