+++ src/components/Game/Game.module.scss
@@ -1,5 +1,3 @@
-@import '../../styles/colors.scss';
-
 .game {
   display: flex;
   margin: 40px auto 10px auto;
@@ -7,47 +5,4 @@
   position: relative;
   width: 95%;
   z-index: 0;
-
-  .textDisplay {
-    align-self: flex-end;
-    height: 90%;
-    margin-bottom: 80px;
-    overflow-y: auto;
-    width: 100%;
-
-    .messageItem {
-      align-items: flex-start;
-      display: flex;
-      margin: 10px auto;
-      user-select: none;
-
-      .character {
-        color: cyan;
-        display: inline-block;
-        margin-right: 20px;
-        overflow: hidden;
-        width: 150px;
-      }
-
-      .message {
-        display: inline-block;
-        font-style: italic;
-      }
-    }
-  }
-
-  .commandCenter {
-    .command {
-      border-top: 1px solid $gray-80;
-      bottom: 10px;
-      left: 0;
-      position: absolute;
-      width: 90%;
-
-      .commandInput {
-        text-align: left;
-        width: 100%;
-      }
-    }
-  }
 }
