+++ src/components/Game/container.js
@@ -3,7 +3,7 @@ import { compose } from 'ramda';
 import withSizes from 'react-sizes';
 import { getPathname } from '../../selectors/routing';
 import { getMessages, getShouldUpdateGame } from '../../selectors/game';
-import { addPlayerInput, loadGame } from '../../actions/game';
+import { loadGame } from '../../actions/game';
 import { connectClient } from '../../actions/pusher';
 import Game from './Game';
 
@@ -16,7 +16,6 @@ const mapStateToProps = state => ({
 const mapSizesToProps = ({ height }) => ({ height });
 
 const mapDispatchToProps = dispatch => ({
-  addPlayerInput: message => dispatch(addPlayerInput(message)),
   connectClient: podId => dispatch(connectClient(podId)),
   loadGame: gameId => dispatch(loadGame(gameId))
 });
